Move the JavaScript relations to lib/relations/JavaScript and register them in AssetGraph.js
