+++ lib/AssetGraph.js
@@ -846,9 +846,27 @@ for (const fileName of fs.readdirSync(
 for (const fileName of fs.readdirSync(
   pathModule.resolve(__dirname, 'relations')
 )) {
-  if (/\.js$/.test(fileName) && fileName !== 'index.js') {
+  if (
+    /\.js$/.test(fileName) &&
+    fileName !== 'index.js' &&
+    fileName !== 'JavaScript'
+  ) {
     AssetGraph.registerRelation(
       pathModule.resolve(__dirname, 'relations', fileName)
     );
   }
+}
+
+const javaScriptRelationsDir = pathModule.resolve(
+  __dirname,
+  'relations',
+  'JavaScript'
+);
+for (const fileName of fs.readdirSync(javaScriptRelationsDir)) {
+  const Relation = require(pathModule.resolve(
+    javaScriptRelationsDir,
+    fileName
+  ));
+  AssetGraph.registerRelation(Relation);
+  AssetGraph.JavaScript.registerRelation(Relation);
 }
