+++ lib/assets/JavaScript.js
@@ -409,19 +409,4 @@ JavaScript.shouldCommentNodeBePreservedInNonPrettyPrintedOutput = comment => {
   );
 };
 
-JavaScript.registerRelation(require('../relations/JavaScriptStaticUrl'));
-JavaScript.registerRelation(
-  require('../relations/JavaScriptServiceWorkerRegistration')
-);
-JavaScript.registerRelation(require('../relations/JavaScriptWorklet'));
-JavaScript.registerRelation(require('../relations/JavaScriptDynamicImport'));
-JavaScript.registerRelation(require('../relations/JavaScriptImportScripts'));
-JavaScript.registerRelation(require('../relations/JavaScriptImport'));
-JavaScript.registerRelation(require('../relations/JavaScriptFetch'));
-JavaScript.registerRelation(require('../relations/JavaScriptWebWorker'));
-JavaScript.registerRelation(require('../relations/JavaScriptExport'));
-JavaScript.registerRelation(require('../relations/JavaScriptSourceUrl'));
-JavaScript.registerRelation(require('../relations/JavaScriptSourceMappingUrl'));
-JavaScript.registerRelation(require('../relations/SystemJsBundle'));
-
 module.exports = JavaScript;
