+++ lib/relations/JavaScript/JavaScriptDynamicImport.js
@@ -1,4 +1,4 @@
-const Relation = require('./Relation');
+const Relation = require('../Relation');
 
 // Avoid modelling "bare" imports as relations (import { uniq } from 'lodash')
 function isImportUrl(href) {
