+++ lib/relations/JavaScript/JavaScriptExport.js
@@ -1,17 +1,17 @@
-const Relation = require('./Relation');
+const Relation = require('../Relation');
 
 // Avoid modelling "bare" imports as relations (import { uniq } from 'lodash')
 function isImportUrl(href) {
   return /^\.{0,2}\/|^[a-z0-9+]+:/.test(href);
 }
 
-class JavaScriptImport extends Relation {
+class JavaScriptExport extends Relation {
   static get selectors() {
     return {
-      ImportDeclaration(node) {
+      ':matches(ExportNamedDeclaration, ExportAllDeclaration)[source]'(node) {
         if (isImportUrl(node.source.value)) {
           return {
-            type: 'JavaScriptImport',
+            type: 'JavaScriptExport',
             node,
             href: node.source.value
           };
@@ -29,14 +29,15 @@ class JavaScriptImport extends Relation {
   }
 
   attach(position, adjacentRelation) {
+    // TODO: Create an ExportNamedDeclaration if specifiers are given?
     this.node = {
-      type: 'ImportDeclaration',
+      type: 'ExportAllDeclaration',
       specifiers: [],
       source: { type: 'Literal', value: '<urlGoesHere>' }
     };
     if (position === 'last') {
       const otherOutgoingJavaScriptImports = this.from.outgoingRelations.filter(
-        relation => relation !== this && relation.type === 'JavaScriptImport'
+        relation => relation !== this && relation.type === 'JavaScriptExport'
       );
       if (otherOutgoingJavaScriptImports.length > 0) {
         position = 'after';
@@ -55,7 +56,7 @@ class JavaScriptImport extends Relation {
       const i = this.from.parseTree.body.indexOf(adjacentRelation.node);
       if (i === -1) {
         throw new Error(
-          'JavaScriptImport#attach: adjacentRelation.node not found in adjacentRelation.from.parseTree.body'
+          'JavaScriptExport#attach: adjacentRelation.node not found in adjacentRelation.from.parseTree.body'
         );
       }
       this.from.parseTree.body.splice(
@@ -71,7 +72,7 @@ class JavaScriptImport extends Relation {
     const i = this.from.parseTree.body.indexOf(this.node);
     if (i === -1) {
       throw new Error(
-        'relations.JavaScriptImport#detach: this.node not found in module array of this.from.parseTree.body.'
+        'relations.JavaScriptExport#detach: this.node not found in this.from.parseTree.body.'
       );
     }
     this.from.parseTree.body.splice(i, 1);
@@ -79,10 +80,10 @@ class JavaScriptImport extends Relation {
   }
 }
 
-Object.assign(JavaScriptImport.prototype, {
-  targetType: 'JavaScript',
+Object.assign(JavaScriptExport.prototype, {
+  nonBareRelative: true,
 
-  nonBareRelative: true
+  targetType: 'JavaScript'
 });
 
-module.exports = JavaScriptImport;
+module.exports = JavaScriptExport;
