+++ lib/relations/JavaScript/JavaScriptFetch.js
@@ -1,5 +1,5 @@
-const replaceDescendantNode = require('../replaceDescendantNode');
-const Relation = require('./Relation');
+const replaceDescendantNode = require('../../replaceDescendantNode');
+const Relation = require('../Relation');
 
 class JavaScriptFetch extends Relation {
   static get selectors() {
