+++ lib/relations/JavaScript/JavaScriptImport.js
@@ -1,17 +1,17 @@
-const Relation = require('./Relation');
+const Relation = require('../Relation');
 
 // Avoid modelling "bare" imports as relations (import { uniq } from 'lodash')
 function isImportUrl(href) {
   return /^\.{0,2}\/|^[a-z0-9+]+:/.test(href);
 }
 
-class JavaScriptExport extends Relation {
+class JavaScriptImport extends Relation {
   static get selectors() {
     return {
-      ':matches(ExportNamedDeclaration, ExportAllDeclaration)[source]'(node) {
+      ImportDeclaration(node) {
         if (isImportUrl(node.source.value)) {
           return {
-            type: 'JavaScriptExport',
+            type: 'JavaScriptImport',
             node,
             href: node.source.value
           };
@@ -29,15 +29,14 @@ class JavaScriptExport extends Relation {
   }
 
   attach(position, adjacentRelation) {
-    // TODO: Create an ExportNamedDeclaration if specifiers are given?
     this.node = {
-      type: 'ExportAllDeclaration',
+      type: 'ImportDeclaration',
       specifiers: [],
       source: { type: 'Literal', value: '<urlGoesHere>' }
     };
     if (position === 'last') {
       const otherOutgoingJavaScriptImports = this.from.outgoingRelations.filter(
-        relation => relation !== this && relation.type === 'JavaScriptExport'
+        relation => relation !== this && relation.type === 'JavaScriptImport'
       );
       if (otherOutgoingJavaScriptImports.length > 0) {
         position = 'after';
@@ -56,7 +55,7 @@ class JavaScriptExport extends Relation {
       const i = this.from.parseTree.body.indexOf(adjacentRelation.node);
       if (i === -1) {
         throw new Error(
-          'JavaScriptExport#attach: adjacentRelation.node not found in adjacentRelation.from.parseTree.body'
+          'JavaScriptImport#attach: adjacentRelation.node not found in adjacentRelation.from.parseTree.body'
         );
       }
       this.from.parseTree.body.splice(
@@ -72,7 +71,7 @@ class JavaScriptExport extends Relation {
     const i = this.from.parseTree.body.indexOf(this.node);
     if (i === -1) {
       throw new Error(
-        'relations.JavaScriptExport#detach: this.node not found in this.from.parseTree.body.'
+        'relations.JavaScriptImport#detach: this.node not found in module array of this.from.parseTree.body.'
       );
     }
     this.from.parseTree.body.splice(i, 1);
@@ -80,10 +79,10 @@ class JavaScriptExport extends Relation {
   }
 }
 
-Object.assign(JavaScriptExport.prototype, {
-  nonBareRelative: true,
+Object.assign(JavaScriptImport.prototype, {
+  targetType: 'JavaScript',
 
-  targetType: 'JavaScript'
+  nonBareRelative: true
 });
 
-module.exports = JavaScriptExport;
+module.exports = JavaScriptImport;
