+++ lib/relations/JavaScript/JavaScriptImportScripts.js
@@ -1,4 +1,4 @@
-const Relation = require('./Relation');
+const Relation = require('../Relation');
 
 class JavaScriptImportScripts extends Relation {
   static get selectors() {
