+++ lib/relations/JavaScript/JavaScriptServiceWorkerRegistration.js
@@ -1,4 +1,4 @@
-const Relation = require('./Relation');
+const Relation = require('../Relation');
 
 class JavaScriptServiceWorkerRegistration extends Relation {
   static get selectors() {
