+++ lib/relations/JavaScript/JavaScriptSourceMappingUrl.js
@@ -1,4 +1,4 @@
-const Relation = require('./Relation');
+const Relation = require('../Relation');
 
 class JavaScriptSourceMappingUrl extends Relation {
   static get selectors() {
