+++ lib/relations/JavaScript/JavaScriptSourceUrl.js
@@ -1,4 +1,4 @@
-const Relation = require('./Relation');
+const Relation = require('../Relation');
 
 class JavaScriptSourceUrl extends Relation {
   static get selectors() {
