+++ lib/relations/JavaScript/JavaScriptStaticUrl.js
@@ -1,5 +1,5 @@
-const replaceDescendantNode = require('../replaceDescendantNode');
-const Relation = require('./Relation');
+const replaceDescendantNode = require('../../replaceDescendantNode');
+const Relation = require('../Relation');
 
 class JavaScriptStaticUrl extends Relation {
   static get selectors() {
