+++ lib/relations/JavaScript/JavaScriptWebWorker.js
@@ -1,4 +1,4 @@
-const Relation = require('./Relation');
+const Relation = require('../Relation');
 
 class JavaScriptWebWorker extends Relation {
   static get selectors() {
