+++ lib/relations/JavaScript/JavaScriptWorklet.js
@@ -1,7 +1,7 @@
-const replaceDescendantNode = require('../replaceDescendantNode');
-const Relation = require('./Relation');
+const replaceDescendantNode = require('../../replaceDescendantNode');
+const Relation = require('../Relation');
 
-const { hrefTypes, getHrefType } = require('../util/getHrefType');
+const { hrefTypes, getHrefType } = require('../../util/getHrefType');
 
 class JavaScriptWorklet extends Relation {
   static get selectors() {
